#1028 证书校验适配新的签名框架
