--- executor/authority/test/secp256r1/chain33.auth.test.toml
@@ -1,88 +1,0 @@
-Title="chain33"
-
-[log]
-# 日志级别，支持debug(dbug)/info/warn/error(eror)/crit
-loglevel = "debug"
-logConsoleLevel = "info"
-# 日志文件名，可带目录，所有生成的日志文件都放到此目录下
-logFile = "logs/chain33.log"
-# 单个日志文件的最大值（单位：兆）
-maxFileSize = 20
-# 最多保存的历史日志文件个数
-maxBackups = 20
-# 最多保存的历史日志消息（单位：天）
-maxAge = 28
-# 日志文件名是否使用本地事件（否则使用UTC时间）
-localTime = true
-# 历史日志文件是否压缩（压缩格式为gz）
-compress = false
-# 是否打印调用源文件和行号
-callerFile = true
-# 是否打印调用方法
-callerFunction = true
-
-[blockchain]
-defCacheSize=512
-maxFetchBlockNum=128
-timeoutSeconds=5
-batchBlockNum=128
-driver="memdb"
-dbPath="datadir"
-dbCache=64
-isStrongConsistency=false
-singleMode=true
-batchsync=false
-isRecordBlockSequence=true
-
-[p2p]
-types=["dht"]
-enable=true
-msgCacheSize=10240
-driver="memdb"
-dbPath="datadir/addrbook"
-dbCache=4
-grpcLogFile="grpc33.log"
-
-
-[rpc]
-jrpcBindAddr="localhost:8801"
-grpcBindAddr="localhost:8802"
-whitelist=["127.0.0.1"]
-jrpcFuncWhitelist=["*"]
-grpcFuncWhitelist=["*"]
-
-[mempool]
-poolCacheSize=10240
-minTxFeeRate=0
-
-[consensus]
-name="solo"
-minerstart=true
-genesis="14KEKbYtKKQm4wMthSK9J4La4nAiidGozt"
-genesisBlockTime=1514533394
-hotkeyAddr="12qyocayNF7Lv6C9qW4avxs2E7U41fKSfv"
-
-[store]
-name="mavl"
-driver="memdb"
-dbPath="datadir/mavltree"
-dbCache=128
-
-[wallet]
-minFee=1000000
-driver="memdb"
-dbPath="datadir/wallet"
-dbCache=16
-signType="secp256k1"
-minerwhitelist=["*"]
-
-[exec]
-enableStat=false
-
-[exec.sub.cert]
-# 是否启用证书验证和签名
-enable=true
-# 加密文件路径
-cryptoPath="./authdir/crypto"
-# 带证书签名类型，支持"auth_ecdsa", "auth_sm2"
-signType="auth_ecdsa"
