--- executor/authority/test/sm2/authdir/crypto/cacerts/ca-cert.pem
@@ -1,13 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIB8zCCAZigAwIBAgIRAPKOX/eTl8vyusy1iUhwTwwwCgYIKoEcz1UBg3UwSjEO
-MAwGA1UEBhMFQ2hpbmExETAPBgNVBAgTCFpoZWppYW5nMREwDwYDVQQHEwhIYW5n
-emhvdTESMBAGA1UEAxMJQ2hhaW4zM0NBMB4XDTIxMDMwODA4NDYwNVoXDTIxMDYx
-NjA4NDYwNVowSjEOMAwGA1UEBhMFQ2hpbmExETAPBgNVBAgTCFpoZWppYW5nMREw
-DwYDVQQHEwhIYW5nemhvdTESMBAGA1UEAxMJQ2hhaW4zM0NBMFkwEwYHKoZIzj0C
-AQYIKoEcz1UBgi0DQgAE/9nvE0lXE0sfRp/LM3C4XXLedsIy0KW7A/92BTwB4qNZ
-RrlKHHcgnYdx8UZxhVv5gPGJOF91SkF9rJTCStHdCaNfMF0wDgYDVR0PAQH/BAQD
-AgGmMA8GA1UdJQQIMAYGBFUdJQAwDwYDVR0TAQH/BAUwAwEB/zApBgNVHQ4EIgQg
-qmYSkbqPt/gDlIT+4NhYB1LuQVI6eaFlqssGgQvVQckwCgYIKoEcz1UBg3UDSQAw
-RgIhAIx+IVbnzahkWOdsjYPuCzW/6uhz5LFC1x5aRlpODsTTAiEA7HdS7+fLz6or
-oNbOYW/LU6q/ioNoQji+gDroBlEoW7M=
------END CERTIFICATE-----
