+++ executor/execenv.go
@@ -7,8 +7,6 @@ package executor
 import (
 	"bytes"
 
-	"github.com/33cn/chain33/executor/authority"
-
 	"github.com/33cn/chain33/account"
 	"github.com/33cn/chain33/client"
 	"github.com/33cn/chain33/client/api"
@@ -228,15 +226,6 @@ func (e *executor) execCheckTx(tx *types.Transaction, index int) error {
 			return types.ErrBalanceLessThanTenTimesFee
 		}
 
-	}
-
-	// 证书检查
-	if authority.IsAuthEnable && e.cfg.IsFork(e.height, "ForkExecCert") {
-		err := authority.Author.Validate(tx.GetSignature())
-		if err != nil {
-			elog.Error("execCheckTx.authority", "exec", string(tx.Execer), "err", err.Error())
-			return err
-		}
 	}
 
 	return exec.CheckTx(tx, index)
