+++ system/crypto/secp256r1/secp256r1_test.go
@@ -1,4 +1,4 @@
-package ecdsa
+package secp256r1
 
 import (
 	"strings"
