+++ system/crypto/secp256r1/utils.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package ecdsa
+package secp256r1
 
 import (
 	"crypto/ecdsa"
