+++ system/crypto/sm2/test/authdir/crypto/intermediatecerts/org1-cert.pem
@@ -1,0 +1,13 @@
+-----BEGIN CERTIFICATE-----
+MIIB8jCCAZigAwIBAgIRAMMknvJdm6yI3c7je/fqzKswCgYIKoEcz1UBg3UwSjEO
+MAwGA1UEBhMFQ2hpbmExETAPBgNVBAgTCFpoZWppYW5nMREwDwYDVQQHEwhIYW5n
+emhvdTESMBAGA1UEAxMJQ2hhaW4zM0NBMB4XDTIxMDMwODA4NDYwNVoXDTIxMDYx
+NjA4NDYwNVowSjEOMAwGA1UEBhMFQ2hpbmExETAPBgNVBAgTCFpoZWppYW5nMREw
+DwYDVQQHEwhIYW5nemhvdTESMBAGA1UEAxMJQ2hhaW4zM0NBMFkwEwYHKoZIzj0C
+AQYIKoEcz1UBgi0DQgAESaDx1yvCU8tZXvtRh7RlaAR0ljnfbGAem4I/iSTNjeNJ
+K6UP4/mS+fYNJuR+ogG2LJfxCrVIuMnghRVNoRisV6NfMF0wDgYDVR0PAQH/BAQD
+AgGmMA8GA1UdJQQIMAYGBFUdJQAwDwYDVR0TAQH/BAUwAwEB/zApBgNVHQ4EIgQg
+BXnhUjdhQyI6cEkZYatUh4Hh2BEgczot+rsHpWI/9gAwCgYIKoEcz1UBg3UDSAAw
+RQIgWoXOU6T1yYb5PfGqUG/R+UXFyB2ulYit/FCX9J26wfwCIQDj1Y6B78gA9go3
+iQ4a/YQ2m+F6ErqI1VLvMc37YzKwbw==
+-----END CERTIFICATE-----
