+++ system/crypto/sm2/test/authdir/crypto/signcerts/user1@org1-cert.pem
@@ -1,0 +1,12 @@
+-----BEGIN CERTIFICATE-----
+MIIB3DCCAYKgAwIBAgIQBjdghSKOJoFdU6sBNJ1/PjAKBggqgRzPVQGDdTBKMQ4w
+DAYDVQQGEwVDaGluYTERMA8GA1UECBMIWmhlamlhbmcxETAPBgNVBAcTCEhhbmd6
+aG91MRIwEAYDVQQDEwlDaGFpbjMzQ0EwHhcNMjEwMzA4MDg0NjA1WhcNMjEwNjE2
+MDg0NjA1WjBHMQ4wDAYDVQQGEwVDaGluYTERMA8GA1UECBMIWmhlamlhbmcxETAP
+BgNVBAcTCEhhbmd6aG91MQ8wDQYDVQQDEwZvcmcxQ0EwWTATBgcqhkjOPQIBBggq
+gRzPVQGCLQNCAASKBToYuZehM7jfCFcc8lcLgBYd0wIyeSaJlLScSuPoZ7ATd1tX
+jqNcbvGlxJpH51rltLgnnAzFB/OpO062psXNo00wSzAOBgNVHQ8BAf8EBAMCB4Aw
+DAYDVR0TAQH/BAIwADArBgNVHSMEJDAigCAFeeFSN2FDIjpwSRlhq1SHgeHYESBz
+Oi36uwelYj/2ADAKBggqgRzPVQGDdQNIADBFAiEA9ejRhk+vUltIBbftlqBxB4z8
+60F+KgcUTP2thE3mhNQCIATwHYAtI5sUB72CJ2oUjY9qhYurjCSi5HOz1eIGkzo5
+-----END CERTIFICATE-----
