+++ types/proto/executor.proto
@@ -79,9 +79,3 @@ message HistoryCertStore {
     int64          curHeigth         = 4;
     int64          nxtHeight         = 5;
 }
-
-message AuthorityCfg {
-    bool   enable     = 1;
-    string cryptoPath = 2;
-    string signType   = 3;
-}
